@@ -23,7 +23,7 @@ module NodeInterfaceTests
 
     element = subject.element("#blahblah")
 
-    capybara_node = stub(:allow_reload!)
+    capybara_node = stub(allow_reload!: nil)
     capybara_stub.session.expects(:first).with("#blahblah", minimum: 0).returns(capybara_node)
     verify_element(element, AePageObjects::Element, subject, capybara_node)
   end
@@ -31,9 +31,9 @@ module NodeInterfaceTests
   def test_element_factory__locator
     subject = node_for_node_tests
 
-    element = subject.element(:locator => ["#blahblah", {:visible => true}])
+    element = subject.element(:locator => ["#blahblah", { visible: true }])
 
-    capybara_node = stub(:allow_reload!)
+    capybara_node = stub(allow_reload!: nil)
     capybara_stub.session.expects(:first).with("#blahblah", visible: true, minimum: 0).returns(capybara_node)
     verify_element(element, AePageObjects::Element, subject, capybara_node)
   end
@@ -43,9 +43,9 @@ module NodeInterfaceTests
 
     subject = node_for_node_tests
 
-    element = subject.element(:locator => ["#blahblah", {:visible => true}], :is => element_class)
+    element = subject.element(:locator => ["#blahblah", { visible: true }], is: element_class)
 
-    capybara_node = stub(:allow_reload!)
+    capybara_node = stub(allow_reload!: nil)
     capybara_stub.session.expects(:first).with("#blahblah", visible: true, minimum: 0).returns(capybara_node)
     verify_element(element, element_class, subject, capybara_node)
   end
