@@ -13,7 +13,7 @@ module AePageObjects
       parent = mock
       parent.stubs(:node).returns(parent_node)
 
-      magazine_node = mock(:allow_reload!)
+      magazine_node = mock(allow_reload!: nil)
       parent_node.expects(:first).with("#18_holder", minimum: 0).returns(magazine_node)
 
       magazine = clip.new(parent, :name => "18_holder", :item_locator => ".some_class")
@@ -31,7 +31,7 @@ module AePageObjects
       parent = mock
       parent.stubs(:node).returns(parent_node)
 
-      magazine_node = mock(:allow_reload!)
+      magazine_node = mock(allow_reload!: nil)
       parent_node.expects(:first).with("#18_holder", minimum: 0).returns(magazine_node)
 
       magazine = clip.new(parent, :name => "18_holder", :item_locator => [:xpath, ".//div[text()='Example Text']"])
@@ -49,7 +49,7 @@ module AePageObjects
       parent = mock
       parent.stubs(:node).returns(parent_node)
 
-      magazine_node = mock(:allow_reload!)
+      magazine_node = mock(allow_reload!: nil)
       parent_node.expects(:first).with("#18_holder", minimum: 0).returns(magazine_node)
 
       magazine = clip.new(parent, :name => "18_holder", :item_locator => [
@@ -70,7 +70,7 @@ module AePageObjects
       parent = mock
       parent.stubs(:node).returns(parent_node)
 
-      magazine_node = mock(:allow_reload!)
+      magazine_node = mock(allow_reload!: nil)
       parent_node.expects(:first).with("#18_holder", minimum: 0).returns(magazine_node)
 
       magazine = clip.new(parent, :name => "18_holder", :item_locator => [
@@ -88,15 +88,15 @@ module AePageObjects
         self.item_class = bullets
       end
 
-      magazine_node = mock(:allow_reload!)
+      magazine_node = mock(allow_reload!: nil)
       parent_node = mock(:first => magazine_node)
       parent = mock(:node => parent_node)
 
       magazine = clip.new(parent, :name => "18_holder", :item_locator => ".some_class")
       magazine.stubs(:item_xpath).returns('item_xpath')
 
-      bullet1_stub = mock(:allow_reload!)
-      magazine_node.expects(:all).with(:xpath, 'item_xpath', {}).returns([bullet1_stub])
+      bullet1_stub = mock(allow_reload!: nil)
+      magazine_node.expects(:all).with(:xpath, 'item_xpath', **{}).returns([bullet1_stub])
       magazine_node.expects(:first).with(:xpath, '(item_xpath)[1]', minimum: 0).returns(bullet1_stub)
       assert_equal bullet1_stub, magazine.at(0).node
     end
@@ -107,16 +107,16 @@ module AePageObjects
         self.item_class = bullets
       end
 
-      magazine_node = mock(:allow_reload!)
+      magazine_node = mock(allow_reload!: nil)
       parent_node = mock(:first => magazine_node)
       parent = mock(:node => parent_node)
 
       magazine = clip.new(parent, :name => "18_holder", :item_locator => [".some_class", { :capybara => 'options' }])
       magazine.stubs(:item_xpath).returns("item_xpath")
 
-      bullet1_stub = mock(:allow_reload!)
-      magazine_node.expects(:all).with(:xpath, 'item_xpath', { capybara: 'options' }).returns([bullet1_stub])
-      magazine_node.expects(:first).with(:xpath, "(item_xpath)[1]", { capybara: 'options', minimum: 0 }).returns(bullet1_stub)
+      bullet1_stub = mock(allow_reload!: nil)
+      magazine_node.expects(:all).with(:xpath, 'item_xpath', capybara: 'options').returns([bullet1_stub])
+      magazine_node.expects(:first).with(:xpath, "(item_xpath)[1]", capybara: 'options', minimum: 0).returns(bullet1_stub)
       assert_equal bullet1_stub, magazine.at(0).node
     end
 
@@ -130,13 +130,13 @@ module AePageObjects
       parent = mock
       parent.stubs(:node).returns(parent_node)
 
-      magazine_node = mock(:allow_reload!)
+      magazine_node = mock(allow_reload!: nil)
       parent_node.expects(:first).with("#18_holder", minimum: 0).returns(magazine_node)
 
       magazine = clip.new(parent, :name => "18_holder")
       magazine.stubs(:item_xpath).returns("item_xpath")
 
-      magazine_node.stubs(:all).with(:xpath, "item_xpath", {}).returns([])
+      magazine_node.stubs(:all).with(:xpath, "item_xpath", **{}).returns([])
 
       assert_equal 0, magazine.size
 
@@ -161,15 +161,15 @@ module AePageObjects
       parent = mock
       parent.stubs(:node).returns(parent_node)
 
-      magazine_node = stub(:allow_reload!)
+      magazine_node = stub(allow_reload!: nil)
       parent_node.expects(:first).with("#18_holder", minimum: 0).returns(magazine_node)
 
       magazine = clip.new(parent, :name => "18_holder")
       magazine.stubs(:item_xpath).returns("item_xpath")
 
-      bullet1_stub = stub(:allow_reload!)
-      bullet2_stub = stub(:allow_reload!)
-      magazine_node.stubs(:all).with(:xpath, "item_xpath", {}).returns([bullet1_stub, bullet2_stub])
+      bullet1_stub = stub(allow_reload!: nil)
+      bullet2_stub = stub(allow_reload!: nil)
+      magazine_node.stubs(:all).with(:xpath, "item_xpath", **{}).returns([bullet1_stub, bullet2_stub])
 
       assert_equal 2, magazine.size
 
