@@ -18,21 +18,21 @@ module AePageObjects
 
         jon = kitty.new
 
-        previous_owners_page_object = stub(:allow_reload!)
+        previous_owners_page_object = stub(allow_reload!: nil)
         capybara_stub.session.expects(:first).with("#previous_owners", minimum: 0).returns(previous_owners_page_object)
 
         previous_owners = verify_element_on_parent_with_intermediary_class(jon, :previous_owners, AePageObjects::Collection, previous_owners_page_object)
 
-        first_owner_page_object = stub(:allow_reload!)
-        previous_owners_page_object.expects(:all).with(:xpath,  ".//*", {}).returns([first_owner_page_object])
+        first_owner_page_object = stub(allow_reload!: nil)
+        previous_owners_page_object.expects(:all).with(:xpath,  ".//*", **{}).returns([first_owner_page_object])
         previous_owners_page_object.expects(:first).with(:xpath, "(.//*)[1]", minimum: 0).returns(first_owner_page_object)
         first_owner = verify_item_element_with_intermediary_class(previous_owners, 0, AePageObjects::Element, first_owner_page_object)
 
-        owner_name_page_object = stub(:allow_reload!)
+        owner_name_page_object = stub(allow_reload!: nil)
         first_owner_page_object.expects(:first).with("#previous_owners_0_owner_name", minimum: 0).returns(owner_name_page_object)
         verify_element_on_parent(first_owner, :owner_name, AePageObjects::Element, owner_name_page_object)
 
-        kitty_name_during_ownership_page_object = stub(:allow_reload!)
+        kitty_name_during_ownership_page_object = stub(allow_reload!: nil)
         first_owner_page_object.expects(:first).with("#previous_owners_0_kitty_name_during_ownership", minimum: 0).returns(kitty_name_during_ownership_page_object)
         verify_element_on_parent(first_owner, :kitty_name_during_ownership, AePageObjects::Element, kitty_name_during_ownership_page_object)
       end
@@ -54,22 +54,22 @@ module AePageObjects
 
         jon = kitty.new
 
-        previous_owners_page_object = stub(:allow_reload!)
+        previous_owners_page_object = stub(allow_reload!: nil)
         capybara_stub.session.expects(:first).with("#previous_owners", minimum: 0).returns(previous_owners_page_object)
 
         previous_owners = verify_element_on_parent_with_intermediary_class(jon, :previous_owners, previous_owners_class, previous_owners_page_object)
 
-        first_owner_page_object = stub(:allow_reload!)
+        first_owner_page_object = stub(allow_reload!: nil)
 
-        previous_owners_page_object.expects(:all).with(:xpath, ".//*", {}).returns([first_owner_page_object])
+        previous_owners_page_object.expects(:all).with(:xpath, ".//*", **{}).returns([first_owner_page_object])
         previous_owners_page_object.expects(:first).with(:xpath, "(.//*)[1]", minimum: 0).returns(first_owner_page_object)
         first_owner = verify_item_element_with_intermediary_class(previous_owners, 0, previous_owners_class.item_class, first_owner_page_object)
 
-        owner_name_page_object = stub(:allow_reload!)
+        owner_name_page_object = stub(allow_reload!: nil)
         first_owner_page_object.expects(:first).with("#previous_owners_0_owner_name", minimum: 0).returns(owner_name_page_object)
         verify_element_on_parent(first_owner, :owner_name, AePageObjects::Element, owner_name_page_object)
 
-        kitty_name_during_ownership_page_object = stub(:allow_reload!)
+        kitty_name_during_ownership_page_object = stub(allow_reload!: nil)
         first_owner_page_object.expects(:first).with("#previous_owners_0_kitty_name_during_ownership", minimum: 0).returns(kitty_name_during_ownership_page_object)
         verify_element_on_parent(first_owner, :kitty_name_during_ownership, AePageObjects::Element, kitty_name_during_ownership_page_object)
       end
@@ -90,22 +90,22 @@ module AePageObjects
 
         jon = kitty.new
 
-        previous_owners_page_object = stub(:allow_reload!)
+        previous_owners_page_object = stub(allow_reload!: nil)
         capybara_stub.session.expects(:first).with("#previous_owners", minimum: 0).returns(previous_owners_page_object)
 
         previous_owners = verify_element_on_parent_with_intermediary_class(jon, :previous_owners, previous_owners_class, previous_owners_page_object)
 
-        first_owner_page_object = stub(:allow_reload!)
+        first_owner_page_object = stub(allow_reload!: nil)
 
-        previous_owners_page_object.expects(:all).with(:xpath, ".//*", {}).returns([first_owner_page_object])
+        previous_owners_page_object.expects(:all).with(:xpath, ".//*", **{}).returns([first_owner_page_object])
         previous_owners_page_object.expects(:first).with(:xpath, "(.//*)[1]", minimum: 0).returns(first_owner_page_object)
         first_owner = verify_item_element_with_intermediary_class(previous_owners, 0, previous_owners_class.item_class, first_owner_page_object)
 
-        owner_name_page_object = stub(:allow_reload!)
+        owner_name_page_object = stub(allow_reload!: nil)
         first_owner_page_object.expects(:first).with("#previous_owners_0_owner_name", minimum: 0).returns(owner_name_page_object)
         verify_element_on_parent(first_owner, :owner_name, AePageObjects::Element, owner_name_page_object)
 
-        kitty_name_during_ownership_page_object = stub(:allow_reload!)
+        kitty_name_during_ownership_page_object = stub(allow_reload!: nil)
         first_owner_page_object.expects(:first).with("#previous_owners_0_kitty_name_during_ownership", minimum: 0).returns(kitty_name_during_ownership_page_object)
         verify_element_on_parent(first_owner, :kitty_name_during_ownership, AePageObjects::Element, kitty_name_during_ownership_page_object)
       end
@@ -129,21 +129,21 @@ module AePageObjects
 
         jon = kitty.new
 
-        previous_owners_page_object = stub(:allow_reload!)
+        previous_owners_page_object = stub(allow_reload!: nil)
         capybara_stub.session.expects(:first).with("#previous_owners", minimum: 0).returns(previous_owners_page_object)
 
         previous_owners = verify_element_on_parent(jon, :previous_owners, previous_owners_class, previous_owners_page_object)
 
-        first_owner_page_object = stub(:allow_reload!)
-        previous_owners_page_object.expects(:all).with(:xpath,  ".//*", {}).returns([first_owner_page_object])
+        first_owner_page_object = stub(allow_reload!: nil)
+        previous_owners_page_object.expects(:all).with(:xpath,  ".//*", **{}).returns([first_owner_page_object])
         previous_owners_page_object.expects(:first).with(:xpath, "(.//*)[1]", minimum: 0).returns(first_owner_page_object)
         first_owner = verify_item_element(previous_owners, 0, previous_owner_class, first_owner_page_object)
 
-        owner_name_page_object = stub(:allow_reload!)
+        owner_name_page_object = stub(allow_reload!: nil)
         first_owner_page_object.expects(:first).with("#previous_owners_0_owner_name", minimum: 0).returns(owner_name_page_object)
         verify_element_on_parent(first_owner, :owner_name, AePageObjects::Element, owner_name_page_object)
 
-        kitty_name_during_ownership_page_object = stub(:allow_reload!)
+        kitty_name_during_ownership_page_object = stub(allow_reload!: nil)
         first_owner_page_object.expects(:first).with("#previous_owners_0_kitty_name_during_ownership", minimum: 0).returns(kitty_name_during_ownership_page_object)
         verify_element_on_parent(first_owner, :kitty_name_during_ownership, AePageObjects::Element, kitty_name_during_ownership_page_object)
       end
@@ -167,21 +167,21 @@ module AePageObjects
 
         jon = kitty.new
 
-        previous_owners_page_object = stub(:allow_reload!)
+        previous_owners_page_object = stub(allow_reload!: nil)
         capybara_stub.session.expects(:first).with("#previous_owners", minimum: 0).returns(previous_owners_page_object)
 
         previous_owners = verify_element_on_parent(jon, :previous_owners, previous_owners_class, previous_owners_page_object)
 
-        first_owner_page_object = stub(:allow_reload!)
-        previous_owners_page_object.expects(:all).with(:xpath,  ".//*", {}).returns([first_owner_page_object])
+        first_owner_page_object = stub(allow_reload!: nil)
+        previous_owners_page_object.expects(:all).with(:xpath,  ".//*", **{}).returns([first_owner_page_object])
         previous_owners_page_object.expects(:first).with(:xpath, "(.//*)[1]", minimum: 0).returns(first_owner_page_object)
         first_owner = verify_item_element(previous_owners, 0, previous_owner_class, first_owner_page_object)
 
-        owner_name_page_object = stub(:allow_reload!)
+        owner_name_page_object = stub(allow_reload!: nil)
         first_owner_page_object.expects(:first).with("#previous_owners_0_owner_name", minimum: 0).returns(owner_name_page_object)
         verify_element_on_parent(first_owner, :owner_name, AePageObjects::Element, owner_name_page_object)
 
-        kitty_name_during_ownership_page_object = stub(:allow_reload!)
+        kitty_name_during_ownership_page_object = stub(allow_reload!: nil)
         first_owner_page_object.expects(:first).with("#previous_owners_0_kitty_name_during_ownership", minimum: 0).returns(kitty_name_during_ownership_page_object)
         verify_element_on_parent(first_owner, :kitty_name_during_ownership, AePageObjects::Element, kitty_name_during_ownership_page_object)
       end
@@ -203,21 +203,21 @@ module AePageObjects
         stub_current_window
         jon = kitty.new
 
-        previous_owners_page_object = stub(:allow_reload!)
+        previous_owners_page_object = stub(allow_reload!: nil)
         capybara_stub.session.expects(:first).with("#previous_owners", minimum: 0).returns(previous_owners_page_object)
 
         previous_owners = verify_element_on_parent_with_intermediary_class(jon, :previous_owners, previous_owners_class, previous_owners_page_object)
 
-        first_owner_page_object = stub(:allow_reload!)
-        previous_owners_page_object.expects(:all).with(:xpath,  ".//*", {}).returns([first_owner_page_object])
+        first_owner_page_object = stub(allow_reload!: nil)
+        previous_owners_page_object.expects(:all).with(:xpath,  ".//*", **{}).returns([first_owner_page_object])
         previous_owners_page_object.expects(:first).with(:xpath, "(.//*)[1]", minimum: 0).returns(first_owner_page_object)
         first_owner = verify_item_element(previous_owners, 0, previous_owner_class, first_owner_page_object)
 
-        owner_name_page_object = stub(:allow_reload!)
+        owner_name_page_object = stub(allow_reload!: nil)
         first_owner_page_object.expects(:first).with("#previous_owners_0_owner_name", minimum: 0).returns(owner_name_page_object)
         verify_element_on_parent(first_owner, :owner_name, AePageObjects::Element, owner_name_page_object)
 
-        kitty_name_during_ownership_page_object = stub(:allow_reload!)
+        kitty_name_during_ownership_page_object = stub(allow_reload!: nil)
         first_owner_page_object.expects(:first).with("#previous_owners_0_kitty_name_during_ownership", minimum: 0).returns(kitty_name_during_ownership_page_object)
         verify_element_on_parent(first_owner, :kitty_name_during_ownership, AePageObjects::Element, kitty_name_during_ownership_page_object)
       end
@@ -237,21 +237,21 @@ module AePageObjects
         stub_current_window
         jon = kitty.new
 
-        previous_owners_page_object = stub(:allow_reload!)
+        previous_owners_page_object = stub(allow_reload!: nil)
         capybara_stub.session.expects(:first).with("#previous_owners", minimum: 0).returns(previous_owners_page_object)
 
         previous_owners = verify_element_on_parent_with_intermediary_class(jon, :previous_owners, AePageObjects::Collection, previous_owners_page_object)
 
-        first_owner_page_object = stub(:allow_reload!)
-        previous_owners_page_object.expects(:all).with(:xpath,  ".//*", {}).returns([first_owner_page_object])
+        first_owner_page_object = stub(allow_reload!: nil)
+        previous_owners_page_object.expects(:all).with(:xpath,  ".//*", **{}).returns([first_owner_page_object])
         previous_owners_page_object.expects(:first).with(:xpath, "(.//*)[1]", minimum: 0).returns(first_owner_page_object)
         first_owner = verify_item_element(previous_owners, 0, previous_owner_class, first_owner_page_object)
 
-        owner_name_page_object = stub(:allow_reload!)
+        owner_name_page_object = stub(allow_reload!: nil)
         first_owner_page_object.expects(:first).with("#previous_owners_0_owner_name", minimum: 0).returns(owner_name_page_object)
         verify_element_on_parent(first_owner, :owner_name, AePageObjects::Element, owner_name_page_object)
 
-        kitty_name_during_ownership_page_object = stub(:allow_reload!)
+        kitty_name_during_ownership_page_object = stub(allow_reload!: nil)
         first_owner_page_object.expects(:first).with("#previous_owners_0_kitty_name_during_ownership", minimum: 0).returns(kitty_name_during_ownership_page_object)
         verify_element_on_parent(first_owner, :kitty_name_during_ownership, AePageObjects::Element, kitty_name_during_ownership_page_object)
       end
@@ -273,21 +273,21 @@ module AePageObjects
 
         jon = kitty.new
 
-        previous_owners_page_object = stub(:allow_reload!)
+        previous_owners_page_object = stub(allow_reload!: nil)
         capybara_stub.session.expects(:first).with("#previous_owners", minimum: 0).returns(previous_owners_page_object)
 
         previous_owners = verify_element_on_parent_with_intermediary_class(jon, :previous_owners, AePageObjects::Collection, previous_owners_page_object)
 
-        first_owner_page_object = stub(:allow_reload!)
-        previous_owners_page_object.expects(:all).with(:xpath,  ".//*", {}).returns([first_owner_page_object])
+        first_owner_page_object = stub(allow_reload!: nil)
+        previous_owners_page_object.expects(:all).with(:xpath,  ".//*", **{}).returns([first_owner_page_object])
         previous_owners_page_object.expects(:first).with(:xpath, "(.//*)[1]", minimum: 0).returns(first_owner_page_object)
         first_owner = verify_item_element_with_intermediary_class(previous_owners, 0, previous_owner_class, first_owner_page_object)
 
-        owner_name_page_object = stub(:allow_reload!)
+        owner_name_page_object = stub(allow_reload!: nil)
         first_owner_page_object.expects(:first).with("#previous_owners_0_owner_name", minimum: 0).returns(owner_name_page_object)
         verify_element_on_parent(first_owner, :owner_name, AePageObjects::Element, owner_name_page_object)
 
-        kitty_name_during_ownership_page_object = stub(:allow_reload!)
+        kitty_name_during_ownership_page_object = stub(allow_reload!: nil)
         first_owner_page_object.expects(:first).with("#previous_owners_0_kitty_name_during_ownership", minimum: 0).returns(kitty_name_during_ownership_page_object)
         verify_element_on_parent(first_owner, :kitty_name_during_ownership, AePageObjects::Element, kitty_name_during_ownership_page_object)
       end
@@ -316,21 +316,21 @@ module AePageObjects
 
         jon = kitty.new
 
-        previous_owners_page_object = stub(:allow_reload!)
+        previous_owners_page_object = stub(allow_reload!: nil)
         capybara_stub.session.expects(:first).with("#previous_owners", minimum: 0).returns(previous_owners_page_object)
 
         previous_owners = verify_element_on_parent_with_intermediary_class(jon, :previous_owners, previous_owners_class, previous_owners_page_object)
 
-        first_owner_page_object = stub(:allow_reload!)
-        previous_owners_page_object.expects(:all).with(:xpath,  ".//*", {}).returns([first_owner_page_object])
+        first_owner_page_object = stub(allow_reload!: nil)
+        previous_owners_page_object.expects(:all).with(:xpath,  ".//*", **{}).returns([first_owner_page_object])
         previous_owners_page_object.expects(:first).with(:xpath, "(.//*)[1]", minimum: 0).returns(first_owner_page_object)
         first_owner = verify_item_element_with_intermediary_class(previous_owners, 0, previous_owner_class, first_owner_page_object)
 
-        owner_name_page_object = stub(:allow_reload!)
+        owner_name_page_object = stub(allow_reload!: nil)
         first_owner_page_object.expects(:first).with("#previous_owners_0_owner_name", minimum: 0).returns(owner_name_page_object)
         verify_element_on_parent(first_owner, :owner_name, AePageObjects::Element, owner_name_page_object)
 
-        kitty_name_during_ownership_page_object = stub(:allow_reload!)
+        kitty_name_during_ownership_page_object = stub(allow_reload!: nil)
         first_owner_page_object.expects(:first).with("#previous_owners_0_kitty_name_during_ownership", minimum: 0).returns(kitty_name_during_ownership_page_object)
         verify_element_on_parent(first_owner, :kitty_name_during_ownership, AePageObjects::Element, kitty_name_during_ownership_page_object)
       end
@@ -344,13 +344,13 @@ module AePageObjects
 
         jon = kitty.new
 
-        previous_owners_page_object = stub(:allow_reload!)
+        previous_owners_page_object = stub(allow_reload!: nil)
         capybara_stub.session.expects(:first).with("#previous_owners", minimum: 0).returns(previous_owners_page_object)
 
         previous_owners = verify_element_on_parent_with_intermediary_class(jon, :previous_owners, AePageObjects::Collection, previous_owners_page_object)
 
-        first_owner_page_object = stub(:allow_reload!)
-        previous_owners_page_object.expects(:all).with(:xpath,  ".//*", {}).returns([first_owner_page_object])
+        first_owner_page_object = stub(allow_reload!: nil)
+        previous_owners_page_object.expects(:all).with(:xpath,  ".//*", **{}).returns([first_owner_page_object])
         previous_owners_page_object.expects(:first).with(:xpath, "(.//*)[1]", minimum: 0).returns(first_owner_page_object)
         first_owner = verify_item_element(previous_owners, 0, AePageObjects::Element, first_owner_page_object)
       end
@@ -369,21 +369,21 @@ module AePageObjects
 
         jon = kitty.new
 
-        previous_owners_page_object = stub(:allow_reload!)
+        previous_owners_page_object = stub(allow_reload!: nil)
         capybara_stub.session.expects(:first).with("whatever you want, baby", minimum: 0).returns(previous_owners_page_object)
 
         previous_owners = verify_element_on_parent_with_intermediary_class(jon, :previous_owners, AePageObjects::Collection, previous_owners_page_object)
 
-        first_owner_page_object = stub(:allow_reload!)
-        previous_owners_page_object.expects(:all).with(:xpath,  ".//*", {}).returns([first_owner_page_object])
+        first_owner_page_object = stub(allow_reload!: nil)
+        previous_owners_page_object.expects(:all).with(:xpath,  ".//*", **{}).returns([first_owner_page_object])
         previous_owners_page_object.expects(:first).with(:xpath, "(.//*)[1]", minimum: 0).returns(first_owner_page_object)
         first_owner = verify_item_element_with_intermediary_class(previous_owners, 0, AePageObjects::Element, first_owner_page_object)
 
-        owner_name_page_object = stub(:allow_reload!)
+        owner_name_page_object = stub(allow_reload!: nil)
         first_owner_page_object.expects(:first).with("#previous_owners_0_owner_name", minimum: 0).returns(owner_name_page_object)
         verify_element_on_parent(first_owner, :owner_name, AePageObjects::Element, owner_name_page_object)
 
-        kitty_name_during_ownership_page_object = stub(:allow_reload!)
+        kitty_name_during_ownership_page_object = stub(allow_reload!: nil)
         first_owner_page_object.expects(:first).with("Kitty Name", minimum: 0).returns(kitty_name_during_ownership_page_object)
         verify_element_on_parent(first_owner, :kitty_name_during_ownership, AePageObjects::Element, kitty_name_during_ownership_page_object)
       end
@@ -404,21 +404,21 @@ module AePageObjects
 
         jon.expects(:page_local_context).returns("hello")
 
-        previous_owners_page_object = stub(:allow_reload!)
+        previous_owners_page_object = stub(allow_reload!: nil)
         capybara_stub.session.expects(:first).with("hello", minimum: 0).returns(previous_owners_page_object)
 
         previous_owners = verify_element_on_parent_with_intermediary_class(jon, :previous_owners, AePageObjects::Collection, previous_owners_page_object)
 
-        first_owner_page_object = stub(:allow_reload!)
-        previous_owners_page_object.expects(:all).with(:xpath,  ".//*", {}).returns([first_owner_page_object])
+        first_owner_page_object = stub(allow_reload!: nil)
+        previous_owners_page_object.expects(:all).with(:xpath,  ".//*", **{}).returns([first_owner_page_object])
         previous_owners_page_object.expects(:first).with(:xpath, "(.//*)[1]", minimum: 0).returns(first_owner_page_object)
         first_owner = verify_item_element_with_intermediary_class(previous_owners, 0, AePageObjects::Element, first_owner_page_object)
 
-        owner_name_page_object = stub(:allow_reload!)
+        owner_name_page_object = stub(allow_reload!: nil)
         first_owner_page_object.expects(:first).with("#previous_owners_0_owner_name", minimum: 0).returns(owner_name_page_object)
         verify_element_on_parent(first_owner, :owner_name, AePageObjects::Element, owner_name_page_object)
 
-        kitty_name_during_ownership_page_object = stub(:allow_reload!)
+        kitty_name_during_ownership_page_object = stub(allow_reload!: nil)
         first_owner_page_object.expects(:first).with("Milkshake", minimum: 0).returns(kitty_name_during_ownership_page_object)
 
         first_owner.expects(:page_local_context).returns("Milkshake")
