@@ -14,7 +14,7 @@ module AePageObjects
         stub_current_window
         jon = kitty.new
 
-        kind_page_object = stub(:allow_reload!)
+        kind_page_object = stub(allow_reload!: nil)
         capybara_stub.session.expects(:first).with("#kind", minimum: 0).returns(kind_page_object)
         verify_element_on_parent(jon, :kind, AePageObjects::Element, kind_page_object)
       end
@@ -31,7 +31,7 @@ module AePageObjects
 
         jon = kitty.new
 
-        kind_page_object = stub(:allow_reload!)
+        kind_page_object = stub(allow_reload!: nil)
         capybara_stub.session.expects(:first).with("Kind Homie", minimum: 0).returns(kind_page_object)
         verify_element_on_parent(jon, :kind, AePageObjects::Element, kind_page_object)
       end
@@ -49,7 +49,7 @@ module AePageObjects
         jon = kitty.new
         jon.expects(:page_local_context).returns("hello")
 
-        kind_page_object = stub(:allow_reload!)
+        kind_page_object = stub(allow_reload!: nil)
         capybara_stub.session.expects(:first).with("hello", minimum: 0).returns(kind_page_object)
         verify_element_on_parent(jon, :kind, AePageObjects::Element, kind_page_object)
       end
@@ -66,7 +66,7 @@ module AePageObjects
 
         jon = kitty.new
 
-        kind_page_object = stub(:allow_reload!)
+        kind_page_object = stub(allow_reload!: nil)
         capybara_stub.session.expects(:first).with("#kind", minimum: 0).returns(kind_page_object)
         verify_element_on_parent(jon, :kind, AePageObjects::Select, kind_page_object)
       end
@@ -83,7 +83,7 @@ module AePageObjects
 
         jon = kitty.new
 
-        kind_page_object = stub(:allow_reload!)
+        kind_page_object = stub(allow_reload!: nil)
         capybara_stub.session.expects(:first).with("#kind", minimum: 0).returns(kind_page_object)
         verify_element_on_parent(jon, :kind, AePageObjects::Checkbox, kind_page_object)
       end
@@ -102,7 +102,7 @@ module AePageObjects
 
         jon = kitty.new
 
-        kind_page_object = stub(:allow_reload!)
+        kind_page_object = stub(allow_reload!: nil)
         capybara_stub.session.expects(:first).with("#kind", minimum: 0).returns(kind_page_object)
 
         verify_element_on_parent(jon, :kind, special_widget, kind_page_object)
@@ -122,7 +122,7 @@ module AePageObjects
 
         jon = kitty.new
 
-        kind_page_object = stub(:allow_reload!)
+        kind_page_object = stub(allow_reload!: nil)
         capybara_stub.session.expects(:first).with("As If!", minimum: 0).returns(kind_page_object)
 
         verify_element_on_parent(jon, :kind, special_widget, kind_page_object)
@@ -149,26 +149,26 @@ module AePageObjects
 
         jon = kitty.new
 
-        tail_page_object = stub(:allow_reload!)
+        tail_page_object = stub(allow_reload!: nil)
         capybara_stub.session.expects(:first).with("#tail_attributes", minimum: 0).returns(tail_page_object)
 
         tail = verify_element_on_parent_with_intermediary_class(jon, :tail, AePageObjects::Element, tail_page_object)
 
-        color_page_object = stub(:allow_reload!)
+        color_page_object = stub(allow_reload!: nil)
         tail_page_object.expects(:first).with("#tail_attributes_color", minimum: 0).returns(color_page_object)
         verify_element_on_parent(tail, :color, AePageObjects::Element, color_page_object)
 
-        size_page_object = stub(:allow_reload!)
+        size_page_object = stub(allow_reload!: nil)
         tail_page_object.expects(:first).with("#tail_attributes_size_attributes", minimum: 0).returns(size_page_object)
         size = verify_element_on_parent_with_intermediary_class(tail, :size, AePageObjects::Element, size_page_object)
 
         assert_equal "Growing!", size.grow!
 
-        length_page_object = stub(:allow_reload!)
+        length_page_object = stub(allow_reload!: nil)
         size_page_object.expects(:first).with("#tail_attributes_size_attributes_length", minimum: 0).returns(length_page_object)
         verify_element_on_parent(size, :length, AePageObjects::Element, length_page_object)
 
-        width_page_object = stub(:allow_reload!)
+        width_page_object = stub(allow_reload!: nil)
         size_page_object.expects(:first).with("#tail_attributes_size_attributes_width", minimum: 0).returns(width_page_object)
         verify_element_on_parent(size, :width, AePageObjects::Element, width_page_object)
       end
@@ -196,26 +196,26 @@ module AePageObjects
 
         jon = kitty.new
 
-        tail_page_object = stub(:allow_reload!)
+        tail_page_object = stub(allow_reload!: nil)
         capybara_stub.session.expects(:first).with("#tail_attributes", minimum: 0).returns(tail_page_object)
 
         tail = verify_element_on_parent(jon, :tail, tail_class, tail_page_object)
 
-        color_page_object = stub(:allow_reload!)
+        color_page_object = stub(allow_reload!: nil)
         tail_page_object.expects(:first).with("#tail_attributes_color", minimum: 0).returns(color_page_object)
         verify_element_on_parent(tail, :color, AePageObjects::Element, color_page_object)
 
-        size_page_object = stub(:allow_reload!)
+        size_page_object = stub(allow_reload!: nil)
         tail_page_object.expects(:first).with("#tail_attributes_size_attributes", minimum: 0).returns(size_page_object)
         size = verify_element_on_parent_with_intermediary_class(tail, :size, AePageObjects::Element, size_page_object)
 
         assert_equal "Growing!", size.grow!
 
-        length_page_object = stub(:allow_reload!)
+        length_page_object = stub(allow_reload!: nil)
         size_page_object.expects(:first).with("#tail_attributes_size_attributes_length", minimum: 0).returns(length_page_object)
         verify_element_on_parent(size, :length, AePageObjects::Element, length_page_object)
 
-        width_page_object = stub(:allow_reload!)
+        width_page_object = stub(allow_reload!: nil)
         size_page_object.expects(:first).with("#tail_attributes_size_attributes_width", anything).returns(width_page_object)
         verify_element_on_parent(size, :width, AePageObjects::Element, width_page_object)
       end
@@ -243,26 +243,26 @@ module AePageObjects
 
         jon = kitty.new
 
-        tail_page_object = stub(:allow_reload!)
+        tail_page_object = stub(allow_reload!: nil)
         capybara_stub.session.expects(:first).with("#tail_attributes", minimum: 0).returns(tail_page_object)
 
         tail = verify_element_on_parent_with_intermediary_class(jon, :tail, tail_base_class, tail_page_object)
 
-        color_page_object = stub(:allow_reload!)
+        color_page_object = stub(allow_reload!: nil)
         tail_page_object.expects(:first).with("#tail_attributes_color", minimum: 0).returns(color_page_object)
         verify_element_on_parent(tail, :color, AePageObjects::Element, color_page_object)
 
-        size_page_object = stub(:allow_reload!)
+        size_page_object = stub(allow_reload!: nil)
         tail_page_object.expects(:first).with("#tail_attributes_size_attributes", minimum: 0).returns(size_page_object)
         size = verify_element_on_parent_with_intermediary_class(tail, :size, AePageObjects::Element, size_page_object)
 
         assert_equal "Growing!", size.grow!
 
-        length_page_object = stub(:allow_reload!)
+        length_page_object = stub(allow_reload!: nil)
         size_page_object.expects(:first).with("#tail_attributes_size_attributes_length", minimum: 0).returns(length_page_object)
         verify_element_on_parent(size, :length, AePageObjects::Element, length_page_object)
 
-        width_page_object = stub(:allow_reload!)
+        width_page_object = stub(allow_reload!: nil)
         size_page_object.expects(:first).with("#tail_attributes_size_attributes_width", minimum: 0).returns(width_page_object)
         verify_element_on_parent(size, :width, AePageObjects::Element, width_page_object)
       end
@@ -288,26 +288,26 @@ module AePageObjects
 
         jon = kitty.new
 
-        tail_page_object = stub(:allow_reload!)
+        tail_page_object = stub(allow_reload!: nil)
         capybara_stub.session.expects(:first).with("what ever you want, baby", minimum: 0).returns(tail_page_object)
 
         tail = verify_element_on_parent_with_intermediary_class(jon, :tail, AePageObjects::Element, tail_page_object)
 
-        color_page_object = stub(:allow_reload!)
+        color_page_object = stub(allow_reload!: nil)
         tail_page_object.expects(:first).with("#tail_attributes_color", minimum: 0).returns(color_page_object)
         verify_element_on_parent(tail, :color, AePageObjects::Element, color_page_object)
 
-        size_page_object = stub(:allow_reload!)
+        size_page_object = stub(allow_reload!: nil)
         tail_page_object.expects(:first).with("Size", minimum: 0).returns(size_page_object)
         size = verify_element_on_parent_with_intermediary_class(tail, :size, AePageObjects::Element, size_page_object)
 
         assert_equal "Growing!", size.grow!
 
-        length_page_object = stub(:allow_reload!)
+        length_page_object = stub(allow_reload!: nil)
         size_page_object.expects(:first).with("#tail_attributes_size_attributes_length", minimum: 0).returns(length_page_object)
         verify_element_on_parent(size, :length, AePageObjects::Element, length_page_object)
 
-        width_page_object = stub(:allow_reload!)
+        width_page_object = stub(allow_reload!: nil)
         size_page_object.expects(:first).with("Fatness", minimum: 0).returns(width_page_object)
         verify_element_on_parent(size, :width, AePageObjects::Element, width_page_object)
       end
@@ -333,28 +333,28 @@ module AePageObjects
 
         jon = kitty.new
 
-        tail_page_object = stub(:allow_reload!)
+        tail_page_object = stub(allow_reload!: nil)
         capybara_stub.session.expects(:first).with("#tail_attributes", minimum: 0).returns(tail_page_object)
 
         tail = verify_element_on_parent_with_intermediary_class(jon, :tail, AePageObjects::Element, tail_page_object)
 
-        color_page_object = stub(:allow_reload!)
+        color_page_object = stub(allow_reload!: nil)
         tail_page_object.expects(:first).with("#tail_attributes_color", minimum: 0).returns(color_page_object)
         verify_element_on_parent(tail, :color, AePageObjects::Element, color_page_object)
 
-        size_page_object = stub(:allow_reload!)
+        size_page_object = stub(allow_reload!: nil)
         tail_page_object.expects(:first).with("#tail_attributes_size_attributes", minimum: 0).returns(size_page_object)
         size = verify_element_on_parent_with_intermediary_class(tail, :size, AePageObjects::Element, size_page_object)
 
         assert_equal "Growing!", size.grow!
 
-        length_page_object = stub(:allow_reload!)
+        length_page_object = stub(allow_reload!: nil)
         size_page_object.expects(:first).with("#tail_attributes_size_attributes_length", minimum: 0).returns(length_page_object)
         verify_element_on_parent(size, :length, AePageObjects::Element, length_page_object)
 
         size.expects(:page_local_context).returns("hello")
 
-        width_page_object = stub(:allow_reload!)
+        width_page_object = stub(allow_reload!: nil)
         size_page_object.expects(:first).with("hello", minimum: 0).returns(width_page_object)
         verify_element_on_parent(size, :width, AePageObjects::Element, width_page_object)
       end
