@@ -17,7 +17,7 @@ module AePageObjects
           is_loaded { loaded << 'doggy' }
         end
 
-        capybara_node = stub(:allow_reload!)
+        capybara_node = stub(allow_reload!: nil)
         capybara_stub
         capybara_stub.session.stubs(:first).with('#foo', minimum: 0).returns(capybara_node)
 
