@@ -29,7 +29,7 @@ module AePageObjects
 
       pet           = pet_class.new
 
-      kitty_page_object = stub(:allow_reload!)
+      kitty_page_object = stub(allow_reload!: nil)
       capybara_stub.session.expects(:first).with("#tiger", minimum: 0).returns(kitty_page_object)
       kitty = kitty_class.new(pet, :locator => '#tiger')
 
@@ -48,7 +48,7 @@ module AePageObjects
 
       pet           = pet_class.new
 
-      kitty_page_object = stub(:allow_reload!)
+      kitty_page_object = stub(allow_reload!: nil)
       capybara_stub.session.expects(:first).with("#tiger", minimum: 0).returns(kitty_page_object)
       kitty = kitty_class.new(pet, :name => 'tiger')
 
@@ -67,7 +67,7 @@ module AePageObjects
 
       pet           = pet_class.new
 
-      kitty_page_object = stub(:allow_reload!)
+      kitty_page_object = stub(allow_reload!: nil)
       capybara_stub.session.expects(:first).with("J 2da K", minimum: 0).returns(kitty_page_object)
       kitty = kitty_class.new(pet, :name => 'tiger', :locator => "J 2da K")
 
@@ -96,7 +96,7 @@ module AePageObjects
       kitty_class      = Class.new(AePageObjects::Element)
       kitty_class.any_instance.stubs(:is_loaded?).returns(false)
 
-      kitty_page_object = stub(:allow_reload!)
+      kitty_page_object = stub(allow_reload!: nil)
 
       capybara_stub
       capybara_stub.session.stubs(:first).with('#hi', minimum: 0).returns(kitty_page_object)
@@ -162,7 +162,7 @@ module AePageObjects
 
       pet = pet_class.new
 
-      kitty_capybara_node = stub(:allow_reload!)
+      kitty_capybara_node = stub(allow_reload!: nil)
 
       capybara_stub.session.expects(:first).with("#tiger", minimum: 0).returns(kitty_capybara_node)
       kitty_element = kitty_class.new(pet, :locator => '#tiger')
